fix: refactor lib
